change example code to typescript
