+++ RNFBSDKExample/src/Examples/index.ts
@@ -1,0 +1,2 @@
+export {default as LoginButtonExample} from './LoginButtonExample';
+export {default as ShareDialogExample} from './ShareDialogExample';
